requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.7.1=py36_0
asn1crypto=0.24.0=pypi_0
astor=0.8.0=py36_0
baidu-aip=2.2.13.0=pypi_0
blas=1.0=mkl
blis=0.2.4=pypi_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2019.5.15=1
certifi=2018.11.29=pypi_0
cffi=1.12.3=py36h2e261b9_0
chardet=3.0.4=pypi_0
cryptography=2.7=pypi_0
cuda90=1.0=h6433d27_0
cymem=2.0.2=pypi_0
django=1.11.11=pypi_0
django-cors-headers=2.4.0=pypi_0
django-extensions=2.1.7=pypi_0
django-filter=2.1.0=pypi_0
django-sslserver=0.20=pypi_0
django-werkzeug-debugger-runserver=0.3.1=pypi_0
djangorestframework=3.9.1=pypi_0
drqa=0.1.0=dev_0
elasticsearch=7.0.4=pypi_0
gast=0.2.2=py36_0
google-pasta=0.1.7=py_0
grpcio=1.16.1=py36hf8bcb03_1
h5py=2.9.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
idna=2.8=pypi_0
intel-openmp=2019.4=243
jieba=0.39=pypi_0
joblib=0.13.2=pypi_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
langid=1.1.6=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libprotobuf=3.8.0=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.0.1=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.14=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
murmurhash=1.0.2=pypi_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
nltk=3.4.5=pypi_0
numpy=1.16.0=pypi_0
openssl=1.1.1d=h7b6447c_1
pexpect=4.2.1=pypi_0
pip=19.2.2=py36_0
plac=0.9.6=pypi_0
preshed=2.0.1=pypi_0
prettytable=0.7.2=pypi_0
protobuf=3.8.0=py36he6710b0_0
ptyprocess=0.6.0=pypi_0
pyaudio=0.2.8=pypi_0
pycparser=2.19=py36_0
pyopenssl=19.0.0=pypi_0
pypinyin=0.35.4=pypi_0
python=3.6.9=h265db76_0
pytorch=1.0.0=py3.6_cuda9.0.176_cudnn7.4.1_1
pytz=2018.9=pypi_0
readline=7.0=h7b6447c_5
regex=2019.8.19=pypi_0
requests=2.21.0=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.3.1=pypi_0
setuptools=41.0.1=py36_0
six=1.12.0=pypi_0
spacy=2.1.8=pypi_0
sqlite=3.29.0=h7b6447c_0
sqlparse=0.3.0=pypi_0
srsly=0.1.0=pypi_0
tensorboard=1.14.0=py36hf484d3e_0
tensorflow=1.14.0=mkl_py36h2526735_0
tensorflow-base=1.14.0=mkl_py36h7ce6ba3_0
tensorflow-estimator=1.14.0=py_0
termcolor=1.1.0=pypi_0
thinc=7.0.8=pypi_0
tk=8.6.8=hbc83047_0
tqdm=4.35.0=pypi_0
urllib3=1.24.1=pypi_0
uwsgi=2.0.18=pypi_0
wasabi=0.2.2=pypi_0
websocket-client=0.32.0=pypi_0
werkzeug=0.15.4=pypi_0
wheel=0.33.4=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
